refactor(ui): split attachment availability helpers
